Sync main com develop

See merge request estoque-inteligente/estoque-inteligente-api!20

+++ src/modules/usuario/UsuarioService.ts
@@ -162,7 +162,7 @@ class UsuarioService {
       await this.repository.deletar(userId);
       await mongoose.connection
         .db!.collection('account')
-        .deleteMany({ userId });
+        .deleteMany({ userId: new mongoose.Types.ObjectId(userId) });
       throw error;
     }
